requirements.txt
numpy==1.20.2
pandas==1.2.5
Pillow==8.2.0
scikit-learn==0.24.2
sklearn==0.0
streamlit==0.84.2
